feat(popup): send message to background script when popup is opened
This commit introduces a change where the popup script sends a message to the background script when the popup is opened. This allows the background script to be aware of when the popup is active and perform actions accordingly. A console log was added in the callback to confirm the message was sent.